website/requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
GitPython